+++ lib/pull_down_button.dart
@@ -1,3 +1,15 @@
+// Copyright (c) 2022 https://github.com/notDmDrl
+// Use of this source code is governed by a MIT-style license that can be
+// found in the LICENSE file.
+
+/// Flutter widgets implementing the iOS Pull-Down menu.
+///
+/// To use, import `package:pull_down_button/pull_down_button.dart`
+///
+/// See also:
+///
+/// * https://developer.apple.com/design/human-interface-guidelines/components/menus-and-actions/pull-down-buttons
+///   for Apple guidelines for Pull-Down buttons.
 library pull_down_button;
 
 export 'src/items/actions_row.dart';